# REF/PERF: Move MultiIndex._tuples to MultiIndex._cache