--- a/pandas/tests/indexes/multi/test_compat.py
+++ b/pandas/tests/indexes/multi/test_compat.py
@@ -68,45 +68,54 @@ def test_inplace_mutation_resets_values():
 
     mi1 = MultiIndex(levels=levels, codes=codes)
     mi2 = MultiIndex(levels=levels2, codes=codes)
+
+    assert "_values" not in mi1._cache
+    assert "_values" not in mi2._cache
+
     vals = mi1.values.copy()
     vals2 = mi2.values.copy()
 
-    assert mi1._tuples is not None
+    assert isinstance(mi1._cache["_values"], np.ndarray)
 
     # Make sure level setting works
     new_vals = mi1.set_levels(levels2).values
     tm.assert_almost_equal(vals2, new_vals)
 
-    # Non-inplace doesn't kill _tuples [implementation detail]
-    tm.assert_almost_equal(mi1._tuples, vals)
+    # Non-inplace doesn't drop _values from _cache [implementation detail]
+    tm.assert_almost_equal(mi1._cache["_values"], vals)
 
     # ...and values is still same too
     tm.assert_almost_equal(mi1.values, vals)
 
-    # Inplace should kill _tuples
+    # Inplace should drop _values from _cache
     with tm.assert_produces_warning(FutureWarning):
         mi1.set_levels(levels2, inplace=True)
+    assert "_values" not in mi1._cache
     tm.assert_almost_equal(mi1.values, vals2)
 
     # Make sure label setting works too
     codes2 = [[0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0]]
     exp_values = np.empty((6,), dtype=object)
     exp_values[:] = [(1, "a")] * 6
-
     # Must be 1d array of tuples
     assert exp_values.shape == (6,)
-    new_values = mi2.set_codes(codes2).values
+
+    new_mi = mi2.set_codes(codes2)
+    assert "_values" not in new_mi._cache
+    new_values = new_mi.values
+    assert "_values" in new_mi._cache
 
     # Not inplace shouldn't change
-    tm.assert_almost_equal(mi2._tuples, vals2)
+    tm.assert_almost_equal(mi2._cache["_values"], vals2)
 
     # Should have correct values
     tm.assert_almost_equal(exp_values, new_values)
 
-    # ...and again setting inplace should kill _tuples, etc
-    with tm.assert_produces_warning(FutureWarning):
-        mi2.set_codes(codes2, inplace=True)
+    # ...and again setting inplace should drop _values from _cache, etc
+    mi2.set_codes(codes2, inplace=True)
+    assert "_values" not in mi2._cache
     tm.assert_almost_equal(mi2.values, new_values)
+    assert "_values" in mi2._cache
 
 
 def test_ndarray_compat_properties(idx, compat_props):